Delete Sidebar Component move content to overview component
