+++ src/main/webapp/app/overview/course-overview.component.scss
