--- src/main/webapp/app/shared/sidebar/sidebar.component.html
@@ -1,126 +1,0 @@
-<div class="sidebar-wrapper">
-    <div class="vw-100 bg-body revert-card-margin">
-        <mat-sidenav-container class="vw-100 vh-100" [ngClass]="{ 'sidenav-height-dev': !inProduction }" [ngClass]="{ 'container-closed': isCollapsed }">
-            <mat-sidenav
-                [ngClass]="{ 'sidenav-height-dev': !inProduction }"
-                class="bg-white rounded-end rounded-3"
-                opened="true"
-                mode="side"
-                [fixedInViewport]="true"
-                [fixedTopGap]="87"
-                @slideIn
-            >
-                <div class="sidebar-container d-flex h-100 justify-content-between flex-column">
-                    <!-- NavItems -->
-                    <div>
-                        <a routerLink="overview" class="d-flex p-3 align-items-center text-decoration-none">
-                            @if (course && course.courseIcon) {
-                                <div class="me-3">
-                                    <jhi-secured-image [src]="course.courseIcon" class="rounded-circle" />
-                                </div>
-                            } @else {
-                                <div class="me-2">
-                                    <img src="content/images/missing_image.png" alt="" width="32" height="32" class="rounded-circle" />
-                                </div>
-                            }
-                            @if (!isCollapsed) {
-                                <div class="h6 mb-0 fw-bold text-body">{{ course?.title }}</div>
-                            }
-                        </a>
-                        <hr class="mt-0" />
-                        <ul class="navbar-nav justify-content-end flex-grow-1 text-decoration-none">
-                            <!-- Overview page doesn't exist yet -->
-                            <!-- <li class="nav-item">
-                          <a class="nav-link nav-link-sidebar px-4" aria-current="page" href="#">
-                              <fa-icon [fixedWidth]="true" [icon]="faFileLines" class="ms-2 me-3"> </fa-icon>
-                              Overview</a
-                          >
-                      </li> -->
-
-                            <!-- Todo get Items with Resolver?  -->
-
-                            @for (sidebarItem of sidebarItems; track sidebarItem) {
-                                @if (sidebarItem.hasInOrionProperty && sidebarItem.showInOrionWindow !== undefined) {
-                                    <li class="nav-item">
-                                        <a
-                                            class="nav-link nav-link-sidebar px-3"
-                                            [ngClass]="{ 'guided-tour': sidebarItem.guidedTour }"
-                                            jhiOrionFilter
-                                            [showInOrionWindow]="sidebarItem.showInOrionWindow"
-                                            [routerLink]="sidebarItem.routerLink"
-                                            routerLinkActive="active"
-                                            [jhiFeatureToggleHide]="sidebarItem.featureToggle"
-                                        >
-                                            <ng-template *ngTemplateOutlet="navIconAndText; context: { $implicit: sidebarItem }" />
-                                        </a>
-                                    </li>
-                                } @else {
-                                    <li class="nav-item">
-                                        <a
-                                            class="nav-link nav-link-sidebar px-3"
-                                            [ngClass]="{ 'guided-tour': sidebarItem.guidedTour }"
-                                            [routerLink]="sidebarItem.routerLink"
-                                            routerLinkActive="active"
-                                            [jhiFeatureToggleHide]="sidebarItem.featureToggle"
-                                        >
-                                            <ng-template *ngTemplateOutlet="navIconAndText; context: { $implicit: sidebarItem }" />
-                                        </a>
-                                    </li>
-                                }
-                            }
-                        </ul>
-                    </div>
-                    <!-- Collapse Item -->
-                    <div
-                        class="double-arrow mb-2"
-                        [ngClass]="{ 'menu-closed': isCollapsed }"
-                        [ngbTooltip]="(isCollapsed ? 'Expand' : 'Collapse') + ' Menu'"
-                        (click)="isCollapsed = !isCollapsed"
-                    >
-                        <div [attr.aria-expanded]="!isCollapsed" aria-controls="collapseBasic">
-                            <fa-icon class="me-negative fa-xs" [icon]="faChevronRight" />
-                            <fa-icon class="fa-xs" [icon]="faChevronRight" />
-                        </div>
-                    </div>
-                </div>
-            </mat-sidenav>
-            <mat-sidenav-content [ngClass]="{ 'sidenav-height-dev': !inProduction }">
-                @if (course) {
-                    <div class="h5 bg-white m-3 mt-0 px-3 py-2 rounded rounded-3 d-flex justify-content-between">
-                        <div class="align-self-center">{{ 'artemisApp.courseOverview.menu.' + pageTitle | artemisTranslate }}</div>
-
-                        <div class="d-flex">
-                            @if (isNotManagementView && this.course.isAtLeastTutor) {
-                                <div id="manage-button" clasS="me-3">
-                                    <a class="btn btn-secondary" [routerLink]="['/course-management', this.course.id]">
-                                        {{ 'artemisApp.courseOverview.manage' | artemisTranslate }}
-                                    </a>
-                                </div>
-                            }
-                            <button class="btn btn-primary btn-md" (click)="loadCourse(true)" [disabled]="refreshingCourse">
-                                <fa-icon [icon]="faSync" [spin]="refreshingCourse" />
-                                <span class="d-none d-sm-inline" jhiTranslate="artemisApp.exercise.refresh">Refresh</span>
-                            </button>
-                        </div>
-                    </div>
-                    <div class="course-body-container bg-white mx-3 p-3 rounded rounded-3">
-                        <ng-container class="d-flex ms-auto" #controlsViewContainer />
-                        <router-outlet (activate)="onSubRouteActivate($event)" (deactivate)="onSubRouteDeactivate()" />
-                        <div class="refresh-overlay" [class.active]="refreshingCourse">
-                            <fa-icon size="lg" [icon]="faCircleNotch" [spin]="true" />
-                        </div>
-                    </div>
-                }
-            </mat-sidenav-content>
-        </mat-sidenav-container>
-    </div>
-</div>
-
-<ng-template #navIconAndText let-sidebarItem>
-    @if (sidebarItem.icon) {
-        <fa-icon [fixedWidth]="true" [icon]="sidebarItem.icon" class="ms-2 me-3" />
-    }
-    @if (!isCollapsed) {
-        <span [jhiTranslate]="sidebarItem.translation">{{ sidebarItem.name }}</span>
-    }
-</ng-template>
