--- src/main/webapp/app/shared/sidebar/sidebar.module.ts
@@ -1,36 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ArtemisSharedModule } from 'app/shared/shared.module';
-import { ArtemisSidePanelModule } from 'app/shared/side-panel/side-panel.module';
-import { OrionModule } from 'app/shared/orion/orion.module';
-import { FeatureToggleModule } from 'app/shared/feature-toggle/feature-toggle.module';
-import { ArtemisHeaderExercisePageWithDetailsModule } from 'app/exercises/shared/exercise-headers/exercise-headers.module';
-import { ArtemisComplaintsModule } from 'app/complaints/complaints.module';
-import { ArtemisCoursesRoutingModule } from 'app/overview/courses-routing.module';
-import { ArtemisSharedPipesModule } from 'app/shared/pipes/shared-pipes.module';
-import { RatingModule } from 'app/exercises/shared/rating/rating.module';
-import { ArtemisExerciseButtonsModule } from 'app/overview/exercise-details/exercise-buttons.module';
-import { ArtemisCourseExerciseRowModule } from 'app/overview/course-exercises/course-exercise-row.module';
-import { LineChartModule, NgxChartsModule, PieChartModule } from '@swimlane/ngx-charts';
-import { SidebarComponent } from './sidebar.component';
-
-@NgModule({
-    imports: [
-        ArtemisExerciseButtonsModule,
-        ArtemisCourseExerciseRowModule,
-        ArtemisSharedModule,
-        ArtemisSharedPipesModule,
-        ArtemisSidePanelModule,
-        ArtemisCoursesRoutingModule,
-        ArtemisHeaderExercisePageWithDetailsModule,
-        OrionModule,
-        ArtemisComplaintsModule,
-        FeatureToggleModule,
-        RatingModule,
-        NgxChartsModule,
-        PieChartModule,
-        LineChartModule,
-    ],
-    declarations: [SidebarComponent],
-    exports: [SidebarComponent],
-})
-export class ArtemisSidebarModule {}
